RPI Load optimisation
This version allow to stop Sonos Module update when the module MMM-Sonos is not displayed to the user (because hidden or notification "USER_PRESENCE" = false received from PIS-Sensor for example). 
Immediate update requested when the module is displayed again or user present again. 
No impact for user and no configuration changes. 

Tested on a RPI2 with MM v2.4.1