Provide artBase URL in config to load album art correctly
The underlying sonos API sometimes returns relative urls for album art, which can't be loaded in all circumstances.  This allows you to provide a config option for a base URL for the art to load from.  Put this in config.js:

artBase: 'http://10.107.0.50:1400'

where that is the IP of a Sonos device on your network. It doesn't need to be one in the group playing.